+++ source-code/sdk-js/src/adapter-sveltekit/runtime/server/index.ts
@@ -8,6 +8,7 @@ export {
 	referenceLanguage,
 	languages,
 	getResource,
+	reloadResources,
 	addRuntimeToLocals,
 	getRuntimeFromLocals,
 } from "./state.js"