+++ source-code/sdk-js/src/config/config.ts
@@ -1,4 +1,4 @@
-import { InlangConfig, InlangConfigModule, setupConfig } from "@inlang/core/config"
+import { type InlangConfig, type InlangConfigModule, setupConfig } from "@inlang/core/config"
 import { initialize$import } from '@inlang/core/environment'
 import fs from "node:fs/promises"
 import { resolve } from "node:path"